PINE: Convert from class to namespace
Fewer global objects, indirect includes via headers.
